warning

+++ scilab/modules/graphics/sci_gateway/c/sci_get.c
@@ -90,7 +90,7 @@ int sci_get(char *fname,unsigned long fname_len)
                   return SET_PROPERTY_ERROR;
                 }
              
-             status = GetScreenProperty(getStringMatrixFromStack(stkAdr)[0]);
+             status = GetScreenProperty(getStringMatrixFromStack((int)stkAdr)[0]);
              
              if(status != SET_PROPERTY_SUCCEED) /* Return property */
                {
